+++ app/components/CoinBarChart.tsx
