+++ app/components/CoinForm.tsx
