+++ app/components/CoinLineChart.tsx
