+++ app/components/CoinRow.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import ArrowDown, { ArrowUp } from "../icons/Icons";
+import ArrowDown, { ArrowUp } from "../../icons/Icons";
 import { formatNumber } from "@/app/formatNumber";
 import styled from "styled-components";
 import { useRouter } from "next/navigation";
