+++ app/components/CoinSwiper.tsx
