+++ app/components/Currencies.tsx
