+++ app/components/coin-swiper.css
